refactor(sign-up): merge all the methods into one file
